refactor(web): useCommentsRail lives in hooks/ — components must not import from pages/
layer-order.test.ts caught the upward edge (CommentsRailChrome in
components/ importing the CommentsRail type from pages/, and type-only
edges count). The hook is 'React state over lib', which is the hooks/
layer's definition — moving it down is the rule's own remedy, not an
allowlist entry.
