+++ apps/web/src/pages/DaemonDocumentPage.tsx
@@ -82,7 +82,7 @@ import type { PastDocument } from '../lib/versions-backend.js'
 import { applyViewportRequest } from '../lib/viewport-request.js'
 import { useBrowserToolRegistry } from '../lib/webmcp/use-browser-tool-registry.js'
 import { deriveDaemonPageState } from './daemon-page-state.js'
-import { useCommentsRail } from './use-comments-rail.js'
+import { useCommentsRail } from '../hooks/use-comments-rail.js'
 import { useDaemonDocumentController } from './use-daemon-document-controller.js'
 import { useVersionSaveFlow } from './use-version-save-flow.js'
